fix: address review feedback on chatgpt-import skill
- Simplify SKILL.md flow: ask user to upload ZIP, import immediately
- Remove dry_run — just run the import directly
- Only accept ZIP files (not raw conversations.json)
- Inline parser into skill executor, delete standalone src/import/
- Fix ZIP parser: read compressedSize from central directory to handle
  ZIPs with data descriptors where local header size is 0
