+++ assistant/src/config/vellum-skills/chatgpt-import/SKILL.md
@@ -9,17 +9,15 @@ Import ChatGPT conversation history into Vellum so users can keep their conversa
 
 ## How to guide the user
 
-When a user wants to import their ChatGPT conversations, walk them through this process:
+When a user wants to import their ChatGPT conversations:
 
-1. **Ask for the export file.** Tell the user to go to ChatGPT → Settings → Data controls → Export data. ChatGPT will email them a ZIP file.
-2. **Get the file path.** Ask the user for the path to their downloaded `conversations.json` file or the ZIP file.
-3. **Preview first.** Always run `chatgpt_import` with `dry_run: true` first to show the user what will be imported (conversation count, message count, titles).
-4. **Confirm and import.** After the user confirms, run `chatgpt_import` with `dry_run: false` to perform the actual import.
-5. **Report results.** Tell the user how many conversations and messages were imported, and mention any skipped duplicates.
+1. **Tell the user how to export.** They need to go to ChatGPT → Settings → Data controls → Export data. ChatGPT will email them a ZIP file.
+2. **Ask the user to upload the ZIP file.** Use whatever file upload tools or skills are available to receive the ZIP.
+3. **Run the import.** Once you have the ZIP file path, call `chatgpt_import` with the file path. Report the results — how many conversations and messages were imported, and any skipped duplicates.
 
 ## Notes
 
-- The tool accepts either a `conversations.json` file or the full ZIP export from ChatGPT.
+- Only ZIP files are accepted (the full export archive from ChatGPT).
 - Conversations are deduplicated — re-importing the same file will skip already-imported conversations.
 - Only user and assistant messages are imported (system prompts and tool calls are filtered out).
 - Original timestamps from ChatGPT are preserved.
